test(eve): pair every wire version module with its own contract test
Rule 40's family-level marker check became stale after wire versions moved to append-only vN modules: it ignored session-inbox-wire.v0.ts and .v1.ts entirely while still passing on two marker strings in the facade test.

Require every *-wire.vN.ts to have a colocated *-wire.vN.test.ts instead, and split the contract accordingly: v0 owns frozen legacy fixtures and migration failures; v1 owns the complete schema snapshot, validated encoder, and round-trip. The guard now reflects the file-level protocol history it is supposed to protect.
